Ora funziona anche senza avviare target e watchdog!

+++ src/obstacles.c
@@ -234,6 +234,11 @@ void Positioning(bool array[][MaxWidth], int height, int width)
 
 }
 
+void PositioningEnemy(bool array[][MaxWidth], int x, int y)
+{
+    array[x][y] = 1;
+}
+
 int main(int argc, char *argv[])
 {
     int fd_r_obstacle, fd_w_obstacle;
@@ -241,7 +246,7 @@ int main(int argc, char *argv[])
 
     log_config(FILENAME_LOG, LOG_DEBUG);
     WritePid();
-    kill(watchdogPid, SIG_WRITTEN);
+    if(watchdogPid != -1) kill(watchdogPid, SIG_WRITTEN);
     
     bool obstacle[MaxHeight][MaxWidth];     //Max dimension of the screen
     
@@ -250,16 +255,20 @@ int main(int argc, char *argv[])
     Msg_int msg_int_out;
     Msg_float msg_float_in, msg_float_out;
     int h_Win, w_Win;
+    int enemy_x, enemy_y;
 
-    //Signal from watchdog
-    struct sigaction sa_ping;
-    sa_ping.sa_handler = ping_handler;
-    sa_ping.sa_flags = SA_RESTART;
-    sigemptyset(&sa_ping.sa_mask);
-    
-    if (sigaction(SIG_PING, &sa_ping, NULL) == -1) {
-        perror("Error in ping_handler");
-        exit(EXIT_FAILURE);
+    if(watchdogPid != -1)
+    {
+        //Signal from watchdog
+        struct sigaction sa_ping;
+        sa_ping.sa_handler = ping_handler;
+        sa_ping.sa_flags = SA_RESTART;
+        sigemptyset(&sa_ping.sa_mask);
+        
+        if (sigaction(SIG_PING, &sa_ping, NULL) == -1) {
+            perror("Error in ping_handler");
+            exit(EXIT_FAILURE);
+        }
     }
 
     float Fx, Fy;
@@ -278,7 +287,10 @@ int main(int argc, char *argv[])
                 w_Win = (int)msg_float_in.b;
                 log_debug("Window dimension: %d, %d", h_Win, w_Win);
                 ClearArray(obstacle);
-                Positioning(obstacle, h_Win, w_Win);
+                if(watchdogPid != -1)       //If single player
+                {
+                    Positioning(obstacle, h_Win, w_Win);
+                }
                 write(fd_w_obstacle, obstacle, sizeof(obstacle));
                 break;
             case 'f':       //The BB wants to know the forces that obstacles apply to the drone
@@ -291,6 +303,11 @@ int main(int argc, char *argv[])
                 Set_msg(msg_float_out, 'f', Fx, Fy);
                 write(fd_w_obstacle, &msg_float_out, sizeof(msg_float_out));
                 break;
+            case 'm':       //Multiplayer case: the BB tell us where is the drone of the other player
+                enemy_x = (int)msg_float_in.a;
+                enemy_y = (int)msg_float_in.b;
+                ClearArray(obstacle);
+                PositioningEnemy(obstacle, enemy_x, enemy_y);
             default:
                 log_error("Error: wrong format of the message recived");
                 perror("Format not correct");
